Move commons util tests to commons
